Removed "isLevelLoaded". Instead uses "ballAddress"
The "isLevelLoaded" variable relies on an arbitrary offset of 0x20 that we were not able to see from the "Course" class, so I wanted to remove it. I found a new way of removing loads that only relies on "GameState"

When you press play, the "balls" array will be initialized but it will be empty. When the level loads, the "balls" array will be populated with the player's "Ball4D" or "Balls5D" object. If the "balls" array exists but is empty, we return true in the "isLoading{}".

To help us find if the "balls" array is empty, we have a new variable I called "ballAddress". Basically its just 4 bytes that hold the address to the ball object. If its not zero, then a ball exists. If it is zero, then a ball does not exist. I suppose we could also make this a boolean but either way works.

I had to change a property on the "MemoryWatcher" called the "FailAction", which is a enum with two options:

public enum ReadFailAction
{
    DontUpdate,
    SetZeroOrNull,
}

"DontUpdate" is the default and it means that if the MemoryWatcher fails to read its target (like if a ball object doesn't exist yet), it will return the last read value, which is not good for us. We want "SetZeroOrNull" which will return 0 when the read fails, letting us know that if "ballAddress" is 0, then the ball object does not exist yet and is still loading.